refactor: type 定義を1箇所にまとめ独立させるため
TaskType の定義を TaskType.tsx に切り出し、src/types/ に配置
